Had things apply a bit better when refresh from single page app or full reload.
The observer now attaches to a much higher node (the body!) and things worked out from there.
The non-progressive file list loaded is not targeted more specifically.
Now if a file has comments but the content was too big to load, show a '0' speech bubble to indicate that there are comments there but until the content is loaded we don't know how many.